+++ SimCalorimetry/HGCalSimProducers/interface/HGCDigitizerBase.h
@@ -171,7 +171,7 @@ class HGCDigitizerBase {
   //CUDA specific
   bool isCUDAInit;
   float *h_toa, *h_charge, *d_toa, *d_charge,*devRand;
-  uint8_t *h_type, *d_type;
+  uint16_t *h_type, *d_type;
   uint32_t *h_rawData, *d_rawData;
   curandGenerator_t d_gen;
